requirements.txt
pyyaml>=6.0
pdfplumber>=0.10.0
python-docx>=1.1.0
openpyxl>=3.1.0
html2text>=2020.1.16
markdown>=3.5.0
xhtml2pdf>=0.2.11
